# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MedConsult API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questionnaires:
    get:
      operationId: listQuestionnaires
      tags: [questionnaires]
      summary: List questionnaires
      parameters:
        - name: agencyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Questionnaires list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Questionnaire"
    post:
      operationId: createQuestionnaire
      tags: [questionnaires]
      summary: Create questionnaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionnaireBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Questionnaire"
  /questionnaires/{id}:
    get:
      operationId: getQuestionnaire
      tags: [questionnaires]
      summary: Get questionnaire by ID (includes questions)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questionnaire with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionnaireWithQuestions"
        "404":
          description: Not found
    patch:
      operationId: updateQuestionnaire
      tags: [questionnaires]
      summary: Update questionnaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionnaireBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Questionnaire"
    delete:
      operationId: deleteQuestionnaire
      tags: [questionnaires]
      summary: Delete questionnaire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- QUESTIONS ---
  /questionnaires/{questionnaireId}/questions:
    get:
      operationId: listQuestions
      tags: [questionnaires]
      summary: List questions for a questionnaire
      parameters:
        - name: questionnaireId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questionnaires]
      summary: Add question to questionnaire
      parameters:
        - name: questionnaireId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questionnaires/{questionnaireId}/questions/{id}:
    patch:
      operationId: updateQuestion
      tags: [questionnaires]
      summary: Update question
      parameters:
        - name: questionnaireId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questionnaires]
      summary: Delete question
      parameters:
        - name: questionnaireId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- QUESTIONNAIRE RESPONSES ---
  /questionnaire-responses:
    get:
      operationId: listQuestionnaireResponses
      tags: [questionnaires]
      summary: List questionnaire responses
      parameters:
        - name: appointmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: questionnaireId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Responses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionnaireResponse"
    post:
      operationId: submitQuestionnaireResponse
      tags: [questionnaires]
      summary: Submit questionnaire responses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionnaireResponseBody"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionnaireResponse"
  /questionnaire-responses/{id}:
    get:
      operationId: getQuestionnaireResponse
      tags: [questionnaires]
      summary: Get questionnaire response by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionnaireResponse"
        "404":
          description: Not found
    patch:
      operationId: updateQuestionnaireResponse
      tags: [questionnaires]
      summary: Update questionnaire response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionnaireResponseBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionnaireResponse"

  # --- CONSULTATION RECORDS ---
...
components:
  schemas:
...
    Questionnaire:
      type: object
      properties:
        id:
          type: integer
        agencyId:
          type: integer
        surgeonId:
          type: ["integer", "null"]
        name:
          type: string
        type:
          type: string
          enum: [pre_consultation, surgeon_consultation, declaration]
        isDefault:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - agencyId
        - name
        - type
        - isDefault
        - createdAt
        - updatedAt
    QuestionnaireWithQuestions:
      type: object
      properties:
        id:
          type: integer
        agencyId:
          type: integer
        surgeonId:
          type: ["integer", "null"]
        name:
          type: string
        type:
          type: string
          enum: [pre_consultation, surgeon_consultation, declaration]
        isDefault:
          type: boolean
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - agencyId
        - name
        - type
        - isDefault
        - questions
        - createdAt
        - updatedAt
    CreateQuestionnaireBody:
      type: object
      properties:
        agencyId:
          type: integer
        surgeonId:
          type: ["integer", "null"]
        name:
          type: string
        type:
          type: string
          enum: [pre_consultation, surgeon_consultation, declaration]
        isDefault:
          type: boolean
        copyFromId:
          type: ["integer", "null"]
      required:
        - agencyId
        - name
        - type
    UpdateQuestionnaireBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        isDefault:
          type: ["boolean", "null"]
        surgeonId:
          type: ["integer", "null"]
    Question:
      type: object
      properties:
        id:
          type: integer
        questionnaireId:
          type: integer
        text:
          type: string
        type:
          type: string
          enum: [text, multiple_choice, dropdown, yes_no]
        options:
          type: ["string", "null"]
        isRequired:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - questionnaireId
        - text
        - type
        - isRequired
        - sortOrder
        - createdAt
    CreateQuestionBody:
      type: object
      properties:
        text:
          type: string
        type:
          type: string
          enum: [text, multiple_choice, dropdown, yes_no]
        options:
          type: ["string", "null"]
        isRequired:
          type: boolean
        sortOrder:
          type: integer
      required:
        - text
        - type
    UpdateQuestionBody:
      type: object
      properties:
        text:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
          enum: [text, multiple_choice, dropdown, yes_no, "null"]
        options:
          type: ["string", "null"]
        isRequired:
          type: ["boolean", "null"]
        sortOrder:
          type: ["integer", "null"]
    QuestionnaireResponse:
      type: object
      properties:
        id:
          type: integer
        questionnaireId:
          type: integer
        appointmentId:
          type: ["integer", "null"]
        customerId:
          type: integer
        answers:
          type: string
        submittedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - questionnaireId
        - customerId
        - answers
        - createdAt
        - updatedAt
    CreateQuestionnaireResponseBody:
      type: object
      properties:
        questionnaireId:
          type: integer
        appointmentId:
          type: ["integer", "null"]
        customerId:
          type: integer
        answers:
          type: string
        submittedAt:
          type: ["string", "null"]
      required:
        - questionnaireId
        - customerId
        - answers
    UpdateQuestionnaireResponseBody:
      type: object
      properties:
        answers:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
...